Refactor the code that builds a `Logger` from a bare `Drain` into the new `build::BuilderCommon`.
